+++ packages/devtools_app/lib/src/screens/debugger/controls.dart
@@ -4,7 +4,6 @@
 
 import 'dart:async';
 
-import 'package:codicon/codicon.dart';
 import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_app_shared/utils.dart';
 import 'package:flutter/material.dart' hide Stack;
@@ -90,7 +89,7 @@ class _DebuggingControlsState extends State<DebuggingControls>
       items: [
         ButtonGroupItemData(
           tooltip: 'Pause',
-          icon: Codicons.debugPause,
+          icon: Icons.pause,
           autofocus: true,
           // Disable when paused or selected isolate is a system isolate.
           onPressed:
@@ -100,7 +99,7 @@ class _DebuggingControlsState extends State<DebuggingControls>
         ),
         ButtonGroupItemData(
           tooltip: 'Resume',
-          icon: Codicons.debugContinue,
+          iconAsset: 'icons/debugger/material_resume.png',
           // Enable while paused + not resuming and selected isolate is not
           // a system isolate.
           onPressed:
@@ -117,17 +116,17 @@ class _DebuggingControlsState extends State<DebuggingControls>
       items: [
         ButtonGroupItemData(
           label: 'Step Over',
-          icon: Codicons.debugStepOver,
+          iconAsset: 'icons/debugger/material_step_over.png',
           onPressed: canStep ? () => unawaited(controller.stepOver()) : null,
         ),
         ButtonGroupItemData(
           label: 'Step In',
-          icon: Codicons.debugStepInto,
+          iconAsset: 'icons/debugger/material_step_into.png',
           onPressed: canStep ? () => unawaited(controller.stepIn()) : null,
         ),
         ButtonGroupItemData(
           label: 'Step Out',
-          icon: Codicons.debugStepOut,
+          iconAsset: 'icons/debugger/material_step_out.png',
           onPressed: canStep ? () => unawaited(controller.stepOut()) : null,
         ),
       ],
@@ -181,11 +180,11 @@ class CodeStatisticsControls extends StatelessWidget {
               children: const [
                 _CodeStatsControl(
                   tooltip: 'Show code coverage',
-                  icon: Codicons.checklist,
+                  icon: Icons.checklist,
                 ),
                 _CodeStatsControl(
                   tooltip: 'Show profiler hits',
-                  icon: Codicons.flame,
+                  icon: Icons.local_fire_department,
                 ),
               ],
               onPressed: (index) {